rtc1 0x50 - nrf_timer 2500

+++ btle_hci/advertiser/Include/ts_peripheral.h
@@ -92,7 +92,7 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 void my_timer_start();
 
-uint8_t get_my_timer_time();
+uint16_t get_my_timer_time();
 
 void my_timer_abort();
